refactor component names and shop merchant size
