+++ frontend/src/StyleComponents.js
@@ -12,10 +12,7 @@ export const useStyles = makeStyles(() => ({
     padding: "1rem",
     borderRadius: "1%",
   },
-  merchant: {
-    height: "55%",
-    width: "44%",
-  },
+  merchant: {},
   itemHover: {
     backdropFilter: "brightness(0.3)",
     padding: "1rem",
