+++ frontend/src/components/ShopInventory.js
@@ -2,7 +2,8 @@ import React from "react";
 import PropTypes from "prop-types";
 import { withStyles } from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
-import ItemSlot from "./ItemSlot";
+import ItemSlot from "./Item";
+import DefaultItemIcon from "../images/Icons/DefaultComputerIcon.png";
 
 const styles = (theme) => ({
   root: {
@@ -14,6 +15,14 @@ const styles = (theme) => ({
     textAlign: "center",
     color: theme.palette.text.secondary,
   },
+  item: {
+    flexGrow: 1,
+    backgroundImage: `url(${DefaultItemIcon})`,
+    backgroundSize: "100%",
+    height: "100px",
+    width: "100px",
+    display: "block",
+  },
 });
 
 function FormRow(props) {}
@@ -22,7 +31,7 @@ FormRow.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-function NestedGrid(props) {
+function ShopInventory(props) {
   const { classes } = props;
 
   return (
@@ -51,8 +60,8 @@ function NestedGrid(props) {
   );
 }
 
-NestedGrid.propTypes = {
+ShopInventory.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(NestedGrid);
+export default withStyles(styles)(ShopInventory);
